requirements.txt
pandas
scikit-learn
numpy
seaborn
matplotlib
PyYAML
certifi
dill
joblib
matplotlib
seaborn
numpy
pandas
pyaml
requests
scikit-learn
scipy
statsmodels
pymongo
python-dotenv
dvc
-e .
